+++ src/completion-provider.ts
@@ -23,15 +23,18 @@ export class CompletionProvider<T extends vscode.CompletionItem = vscode.Complet
             parameter.name,
             vscode.CompletionItemKind.Property,
           );
-          completion.documentation = parameter.description
-            ? new vscode.MarkdownString(parameter.description)
-            : undefined;
+          completion.documentation = parseDocumentation(parameter);
           completion.detail = `${parameter.type} ${parameter.isOptional ? '(Optional)' : ''}`;
           if (parameter.type === 'String') {
             completion.insertText = new vscode.SnippetString(`${parameter.name}: '$0'`);
           } else if (parameter.type === 'boolean') {
             completion.insertText = new vscode.SnippetString(
               `${parameter.name}: \${1|true,false|}`,
+            );
+          } else if (parameter.type === 'Enum' && parameter.values.length) {
+            const enumValues = parameter.values.join(',');
+            completion.insertText = new vscode.SnippetString(
+              `${parameter.name}: '\${1|${enumValues}|}'`,
             );
           } else {
             completion.insertText = `${parameter.name}: `;
@@ -43,4 +46,19 @@ export class CompletionProvider<T extends vscode.CompletionItem = vscode.Complet
     }
     return completions;
   }
+}
+
+function parseDocumentation(parameter: {
+  values: string[];
+  description: string;
+}): string | vscode.MarkdownString | undefined {
+  let markdown = '';
+  if (parameter.values.length) {
+    markdown += parameter.values.map(value => `* ${value}\n`).join('');
+    markdown += '\n';
+  }
+  if (parameter.description) {
+    markdown += parameter.description;
+  }
+  return markdown ? new vscode.MarkdownString(markdown) : undefined;
 }