+++ CHANGELOG.md
@@ -5,3 +5,7 @@ Jenkins Doc VSCode extension change log
 ## [1.0.0](https://github.com/Maarti/JenkinsDocExtension/releases/tag/1.0.0) - 2021-03-08
 
 - Initial release
+
+## Next:
+
+- Autocompletion for more than 1400 instructions