+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Jenkins Doc VSCode extension change log
 ## Next:
 
 - Autocompletion for more than 1400 instructions
+- Autocompletion for instructions parameters