+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Jenkins Doc VSCode extension change log
 
+## [1.1.0](https://github.com/Maarti/JenkinsDocExtension/releases/tag/1.1.0) - 2021-03-10
+
+- Autocompletion for more than 1400 instructions
+- Autocompletion for instructions parameters
+- Autocompletion for boolean and enum parameters
+- Improved documentation for nested objects and enum type parameters
+
 ## [1.0.0](https://github.com/Maarti/JenkinsDocExtension/releases/tag/1.0.0) - 2021-03-08
 
 - Initial release